refactor(sdk): Consolidate persistence types into single file
