--- packages/sdk/src/utils/persistence/types.ts
@@ -1,4 +1,0 @@
-export interface Persistence {
-    get(key: string): Promise<string | undefined>
-    set(key: string, value: string): Promise<void>
-}
